moved part of the documentation component builder into html rendering project and wrote base64 component images for html rendering
