+++ components/html-rendering/src/main/resources/org/datacleaner/documentation/template_component.html
